third_party/blink/web_tests/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-public-expected.txt

This is a testharness.js-based test.
FAIL public to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS public to local: missing CORS headers.
PASS public to local: missing PNA header.
FAIL public to local: success. assert_equals: expected "loaded" but got "timeout"
FAIL public to private: failed preflight. assert_equals: expected "timeout" but got "loaded"
PASS public to private: missing CORS headers.
PASS public to private: missing PNA header.
FAIL public to private: success. assert_equals: expected "loaded" but got "timeout"
PASS public to public: no preflight required.
Harness: the test ran to completion.
